added protected event router
